+++ CHANGELOG.md
@@ -13,3 +13,8 @@
 * update README
 * add required Ruby version *>= 1.9.3*
 * add test/spec_help.rb
+
+#### 0.0.1-alpha.3
+
+* add check-smart-status
+* add check-smart